Refactor Connection component
