+++ src/components/connection/Connection.vue
@@ -14,25 +14,7 @@
       label="What is your destination?"
       icon="flight_land"
     )
-    v-layout(row)
-      v-flex(xs6)
-        v-switch(
-          :label="`Remember connection`"
-          v-model="settings.rememberConnection"
-          @change="updateSettings('rememberConnection')"
-        )
-      v-flex(xs6)
-        v-switch(
-          :label="`Auto-Retrieve on next visit`"
-          v-model="settings.autoRetrieveConnection"
-          @change="updateSettings('autoRetrieveConnection')"
-          :disabled="!settings.rememberConnection"
-          )
-    v-switch(
-      label="Submit button on the right"
-      v-model="settings.submitButtonOnRightSide"
-      @change="toggleButtonOrder"
-    )
+    connection-settings
     v-layout(row)
       template(v-for="button in buttons")
         v-flex(
@@ -65,12 +47,14 @@ import { mapGetters } from 'vuex'
 
 import API from '@/api'
 import DepartureSearch from '@/components/departures/DepartureSearch'
+import ConnectionSettings from '@/components/connection/ConnectionSettings'
 import Timeline from '@/components/timeline/Timeline'
 
 export default {
   name: 'Connection',
   components: {
     DepartureSearch,
+    ConnectionSettings,
     Timeline
   },
   mounted () {
@@ -135,16 +119,6 @@ export default {
     }
   },
   methods: {
-    updateSettings (identity) {
-      const payload = {
-        identity: identity,
-        value: this.settings[identity]
-      }
-      this.$store.dispatch('UPDATE_SETTINGS', payload)
-    },
-    toggleButtonOrder () {
-      this.updateSettings('submitButtonOnRightSide')
-    },
     resetForm () {
       this.$store.dispatch('RESET_FORM')
       this.trips = ''
